requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
httpx>=0.24.0
